[studio,plan-build] Set ownership on `./results/` to not be root.
This resolves 2 remaining edge cases when performing Studio builds on
Linux systems with native Studios. In both cases, there is no
pre-existing `./results/` directory:

1. [plan-build] When the pre-build report is written to
`./results/pre_build.env`, attempt to update the ownership of the
`./results/` directory to the same owner/group as the Plan file.
2. [studio] When a log file is being written (default behvaior with `hab
studio build`/`hab pkg build`), attempt to update the ownership of the
`./results/` directory recursively, which at that point will include the
`logs/` directory as well as the target log file.

Closes #3445

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>