[plan-build] Update `_latest_installed_package()` to use `hab pkg path`.
This change deprecates the old implementation of the function which
determines the latest installed package on disk which is used to resolve
dependencies. The original implementation used a combination of `find`,
`sort` (with the `--version-sort` flag), and `head`. The updated
implementation simply uses `hab pkg path` as it provides the same basic
logic. However, the Rust implementation uses the same version string
sorting as the Supervisor and should be considered more correct. As the
GNU version of sort is no longer necessary, the explicit check in
`_find_system_commands()` can be dropped for `sort`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>