Add abilities module
This is the thinnest abstraction over Linux capabilities (mainly to
provide an API we can use on Windows, too). It allows us to check for
the specific ability to perform an operation in order to modify how
the Supervisor behaves.

The capabilities the Supervisor needs to run on Linux with "root-like" powers are:

* `CAP_SETUID` and `CAP_SETGID`: for running supervised processes as
  `SVC_USER`/`SVC_GROUP`.
* `CAP_CHOWN`: to set file and directory permissions to be accessible
  by `SVC_USER`/`SVC_GROUP`.

If the current user does not have those abilities, the Supervisor
simply does not try to perform them. The end result is that the
Supervisor can run as an unprivileged user, and all supervised
processes will run as that same user (`SVC_USER` and `SVC_GROUP` are
effectively ignored).

Windows does not have corresponding facilities, so to preserve
existing behavior, we unconditionally return `true` for all ability
checks.

Signed-off-by: Christopher Maier <cmaier@chef.io>