Allow the Supervisor to run as a non-root user on Linux
When the Supervisor needs to perform an operation that previously
required `root` abilities, we now check for a specific Linux
capability instead. We can then simply not perform those operations,
allowing the Supervisor to run in more varied scenarios, such as
containerized platforms like OpenShift.

At present, the only operations that the Supervisor performs that
require these checks are running supervised processes and service
hooks as `SVC_USER`/`SVC_GROUP`, and granting ownership of certain
files and directories to that same user. If we do not have the
capability to carry out those operations, we simply don't. This
effectively means that `SVC_USER` and `SVC_GROUP` are ignored, and all
supervised processes, as well as the Supervisor itself, will be run by
the same user.

In order to facilitate this, the Launcher protocol had to be changed
to accept UID and GID directly; previously it resolved these data from
names. However, the user the Supervisor is running as is not actually
guaranteed to have a name. As such, we will resolve this information
before invoking the Launcher. The Launcher will use UID/GID
preferentially, but will still accept names to preserve backward
compatibility.

The creation of a service's directory hierarchy was extracted into a
`SvcDir` type, in order to consolidate the creation and capability
checking logic in a single place, and remove it from the main logical
flow of the Manager.

Signed-off-by: Christopher Maier <cmaier@chef.io>