Simplify sup-protocol build
We had a lot of custom code in `build.rs` that was essentially a copy
of core `prost-build` code in order to generate simple `impl` blocks
for each of our core messages.

This code was rather intricate, and it wasn't very obvious what it was
doing. Additionally, it diverges from our (later, arguably better)
pattern in the `butterfly` crate of just including the base generated
types, and writing our own `impl` blocks.

Yes, it's a bit of additional boilerplate, but it's much more obvious
what is happening now. Additionally, the extra build code was making
dealing with Prost far more "intimidating" than it really needs to be.

Signed-off-by: Christopher Maier <cmaier@chef.io>