Ensure that the Launcher continues to reap services during shutdown
As the Supervisor takes on responsibilities for shutting down
services, the Launcher (which remains the parent of those processes)
needs to continue to reap the processes as it waits for the Supervisor
process to shut down.

This will ensure that the Supervisor can proceed in a timely manner to
shut down everything.

Signed-off-by: Christopher Maier <cmaier@chef.io>