Update docs for running as non-root in a container
With the merging of #5738, the Launcher began to manage its own
state (previously, the Supervisor was actually doing this). This
resulted in the need to write to `/hab/launcher` which is fine, but
ran into trouble in the very specific circumstance of a user trying to
run Habitat in a container, but as a non-root user that was also not
in the `root` group. Our documentation shows how to do this by
mounting in properly writable volumes, but was not updated to take
into account the new `/hab/launcher` directory.

Without this mount being present, nothing could start, because the
Launcher was not able to write its own PID file.

This is just a documentation change; no code needs to change at this
time.

Signed-off-by: Christopher Maier <cmaier@chef.io>