[PR Feedback] SpecDir#specs no longer returns Result
This function originally returned a `Result<Vec<ServiceSpec>>`, with
an error if any spec file were unreadable or contained non-UTF-8
characters. We would log an error, however, if the files wouldn't
otherwise parse, and we would continue to generate a `Vec` to return.

The end result is that we'd skip over some kinds of "bad" files, but
not others.

Having to deal with these errors ended up propagating outward,
affecting several functions in the Manager for no good reason.

Additionally, there were two cases where we blindly called `unwrap` on
the output of this function, meaning that the wrong kind of "bad" file
could end up crashing the Supervisor altogether.

As there doesn't appear to be a compelling reason to differentiate
between these different kinds of "bad" files, any error is now logged,
and `SpecDir#specs` always returns a `Vec<ServiceSpec>`.

(In the course of addressing this, I discovered that we end up reading
the spec files each time through the main Supervisor loop, which seems
pretty wasteful; I'm sure there's something more efficient we could be
doing there. As a result, if there's a bad file in the specs
directory, it will be logged a lot, but that's likely better than
crashing the Supervisor.)

Signed-off-by: Christopher Maier <cmaier@chef.io>