[PR Feedback] Initializing a ReconciliationFlag more explicitly
The derived `Default` implementation will create a flag set to
`false`, since this is exactly what the underlying `AtomicBool` does.

However, it's good to be explicit about such things. Now, we have
`ReconciliationFlag::new`, which delegates to `default()`. It is
explicitly documented as returning a `false` flag, though, and now we
have a unit test to prove it.

Signed-off-by: Christopher Maier <cmaier@chef.io>