[PR Feedback] Make the test for reconciliation more explicit
Previously, testing actually modified the flag, which could be
unexpected and potentially confusing for readers, and could be subject
to unintentional breakage with refactorings.

Now, querying and modification are separated. To aid comprehension,
the full condition is now inlined at the call site, rather than hidden
behind a function call.

(If it helps, you can basically think of the `ReconciliationFlag` as a
kind of signalling channel, where the presence of one or more messages
at all is "true", and draining all messages from the channel is
equivalent to "setting it back to false".)

Signed-off-by: Christopher Maier <cmaier@chef.io>